17September

+++ SpringRest/src/main/java/com/springrest/SpringRest/controller/MyController.java
@@ -1,5 +1,6 @@
 package com.springrest.SpringRest.controller;
 
+import java.sql.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.springrest.SpringRest.entity.Student;
@@ -35,7 +37,19 @@ public class MyController {
 		return this.studentService.getStudent();
 	}
 	
-	// For Searching
+	// For Sorting By DOB
+	@GetMapping("/Students/{dob}")
+	public List<Student> getAllSorting(@PathVariable Date dob){
+		return this.studentService.getSorting(dob);
+	}
+	
+	// For Sorting By Joining Date
+		@GetMapping("/Students/{joining}")
+		public List<Student> getAll2Sorting(@PathVariable Date joining){
+			return this.studentService.getSorting2(joining);
+		}
+	
+	// For Editing
 	@GetMapping("/Students/{userId}")
 	public Student getSingleStudent(@PathVariable String userId){
 		return this.studentService.getUser(Integer.parseInt(userId));
@@ -55,14 +69,26 @@ public class MyController {
 	
 	// For Deleting 
 	@DeleteMapping("/Students/{userId}")
-	public ResponseEntity<HttpStatus> deleteUser(@PathVariable String userId)
+	public ResponseEntity<HttpStatus> deleteUser(@PathVariable String userId, @RequestParam(value = "softDelete",required = false) boolean isSoftDelete)
 	{
 		try {
-			this.studentService.deleteUser(Integer.parseInt(userId));
+			if(isSoftDelete)
+			{	
+				this.studentService.deleteUser(Integer.parseInt(userId));
+			}
+			else {
+				this.studentService.userSoftDelete(userId);
+			}
 			return new ResponseEntity<>(HttpStatus.OK);
 		}
 		catch (Exception e) {
 			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 		}
+	}
+	
+	// Finding By Name
+	@GetMapping("/Student/{name}")
+	public List<Student> getInfoByName(@PathVariable String name){
+		return this.studentService.getUserName(name);
 	}
 }
